+++ streaming_integration_test.go
@@ -1,74 +1,108 @@
 package unleash
 
 import (
+	"fmt"
+	"net/http"
+	"net/http/httptest"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/mock"
 )
 
-func TestStreamingModeConfiguration(t *testing.T) {
-	// Use a mock server instead of external URL
-	server := mockSSEServer(`{"events":[{"type":"hydration","eventId":1,"features":[],"segments":[]}]}`, "")
-	defer server.Close()
+// mockSSEServer creates a test server that simulates SSE responses
+// Since Gock doesn't support SSE streaming, we need a real test server for these tests
+func mockSSEServer(hydrationData, updateData string) *httptest.Server {
+	return httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		switch r.URL.Path {
+		case "/client/register":
+			w.WriteHeader(http.StatusOK)
+			w.Write([]byte(`{}`))
+			
+		case "/client/streaming":
+			// Set SSE headers
+			w.Header().Set("Content-Type", "text/event-stream")
+			w.Header().Set("Cache-Control", "no-cache")
+			w.Header().Set("Connection", "keep-alive")
+			
+			// Send hydration event
+			fmt.Fprintf(w, "event: unleash-connected\ndata: %s\n\n", hydrationData)
+			w.(http.Flusher).Flush()
+			
+			// Send update event if provided
+			if updateData != "" {
+				time.Sleep(50 * time.Millisecond)
+				fmt.Fprintf(w, "event: unleash-updated\ndata: %s\n\n", updateData)
+				w.(http.Flusher).Flush()
+			}
+			
+			// Keep connection open briefly to simulate streaming
+			time.Sleep(100 * time.Millisecond)
+			
+		default:
+			w.WriteHeader(http.StatusNotFound)
+		}
+	}))
+}
+
+// TestStreamingMode_UnleashConnectedEvent tests that the client can process an unleash-connected SSE event
+func TestStreamingMode_UnleashConnectedEvent(t *testing.T) {
+	hydrationData := `{"events":[{"type":"hydration","eventId":1,"features":[{"name":"test-feature","enabled":true,"strategies":[{"name":"default"}],"variants":[],"dependencies":null}],"segments":[]}]}`
 	
-	listener := &MockedListener{}
-	listener.On("OnError", mock.Anything).Return()
-	listener.On("OnWarning", mock.Anything).Return()
-	listener.On("OnReady").Return()
-	listener.On("OnRegistered", mock.AnythingOfType("ClientData")).Return()
-	listener.On("OnCount", mock.AnythingOfType("string"), mock.AnythingOfType("bool")).Return()
+	server := mockSSEServer(hydrationData, "")
+	defer server.Close()
+
+	// Configure client with streaming mode
+	client, err := NewClient(
+		WithUrl(server.URL),
+		WithAppName("test-app"),
+		WithInstanceId("test-instance"),
+		WithCustomHeaders(http.Header{"X-API-KEY": []string{"123"}}),
+		WithExperimentalMode(map[string]string{"type": "streaming"}),
+		WithStorage(&DefaultStorage{}),
+		WithDisableMetrics(true),
+	)
+	assert.NoError(t, err)
+	defer client.Close()
+
+	// Wait for the client to connect and process the event
+	client.WaitForReady()
+
+	// Verify the feature is enabled after hydration
+	assert.True(t, client.IsEnabled("test-feature"), "test-feature should be enabled after hydration")
+
+	// Test with a non-existent feature
+	assert.False(t, client.IsEnabled("non-existent-feature"), "non-existent-feature should be disabled")
+}
+
+// TestStreamingMode_UnleashUpdatedEvent tests processing of a simple unleash-updated event
+func TestStreamingMode_UnleashUpdatedEvent(t *testing.T) {
+	hydrationData := `{"events":[{"type":"hydration","eventId":1,"features":[{"name":"feature-1","enabled":false,"strategies":[{"name":"default"}]}],"segments":[]}]}`
+	updateData := `{"events":[{"type":"feature-updated","eventId":2,"feature":{"name":"feature-1","enabled":true,"strategies":[{"name":"default"}]}}]}`
+	
+	server := mockSSEServer(hydrationData, updateData)
+	defer server.Close()
 
 	client, err := NewClient(
 		WithUrl(server.URL),
 		WithAppName("test-app"),
-		WithListener(listener),
-		WithExperimentalMode(map[string]string{
-			"type": "streaming",
-		}),
+		WithInstanceId("test-instance"),
 		WithDisableMetrics(true),
+		WithExperimentalMode(map[string]string{"type": "streaming"}),
+		WithStorage(&DefaultStorage{}),
 	)
-
 	assert.NoError(t, err)
-	assert.NotNil(t, client)
-	
-	assert.True(t, client.options.IsStreamingMode())
-	assert.True(t, client.repository.IsStreaming())
-	
+	defer client.Close()
+
+	// Wait for initial hydration
 	client.WaitForReady()
-	client.Close()
-}
-
-func TestNonStreamingModeByDefault(t *testing.T) {
-	// Simply test that the default configuration doesn't use streaming mode
-	// without trying to connect to any server
-	config := &configOption{}
 	
-	// Check that streaming mode is disabled by default
-	assert.False(t, config.IsStreamingMode())
-	
-	// Test with a client that has an empty experimental mode
-	config2 := &configOption{
-		experimentalMode: map[string]string{},
-	}
-	assert.False(t, config2.IsStreamingMode())
-}
+	// Feature should initially be disabled
+	assert.False(t, client.IsEnabled("feature-1"), "feature-1 should be initially disabled")
 
-func TestExperimentalModeHelpers(t *testing.T) {
-	config := &configOption{
-		experimentalMode: map[string]string{
-			"type": "streaming",
-		},
-	}
-	assert.True(t, config.IsStreamingMode())
+	// Wait for the update event to be processed
+	time.Sleep(200 * time.Millisecond)
 
-	config2 := &configOption{
-		experimentalMode: map[string]string{
-			"type": "polling",
-		},
-	}
-	assert.False(t, config2.IsStreamingMode())
-
-	config3 := &configOption{}
-	assert.False(t, config3.IsStreamingMode())
+	// Feature should now be enabled after update
+	assert.True(t, client.IsEnabled("feature-1"), "feature-1 should be enabled after update")
 }
